+++ source/DSCResources/MSFT_ADOrganizationalUnit/MSFT_ADOrganizationalUnit.psm1
@@ -48,7 +48,7 @@ function Get-TargetResource
 
     try
     {
-        $ou = Get-ADOrganizationalUnit -Filter "Name -eq '$Name'" -SearchBase $Path `
+        $ou = Get-ADOrganizationalUnit -Filter "Name -eq `"$Name`"" -SearchBase $Path `
             -SearchScope OneLevel -Properties ProtectedFromAccidentalDeletion, Description
     }
     catch [Microsoft.ActiveDirectory.Management.ADIdentityNotFoundException]
@@ -173,7 +173,7 @@ function Test-TargetResource
         {
             # Resource should exist
             $propertiesNotInDesiredState = (
-                Compare-ResourcePropertyState -CurrentValue $targetResource -DesiredValues $PSBoundParameters -IgnoreProperties 'Credential' |
+                Compare-ResourcePropertyState -CurrentValue $targetResource -DesiredValues $PSBoundParameters -IgnoreProperties ('Credential', 'RestoreFromRecycleBin') |
                     Where-Object -Property InDesiredState -eq $false)
 
             if ($propertiesNotInDesiredState)